plot_marker

Function to visualize a selected parameter on a dimensionality reduction.

Usage

plot_marker(
  data,
  param,
  order = FALSE,
  title = "adjust the title",
  limX = NULL,
  limY = NULL,
  dim_red,
  dot_size = 0.1,
  apha = 0.2,
  color_discrete = cluster_palette,
  color_gradient = colors,
  remove_guide = FALSE,
  facet_by_variable = FALSE,
  label_clusters = FALSE,
  label_size = 3.5,
  label_color = "black",
  raster = FALSE
)

Arguments

data

This is the input dataframe for the visualization, this part of the code can still be improved but for the moment you have to cbind the dataframe with the informations you want to plot.

param

Parameter to visualize in the plot, this can be either a continous variable or a categorical one, the function will react differently according.

order

Logical if you want to order the dots in the plot, by expression for example. This can help to find small populations of positive cells.

title

Title of the plot.

limX

Limits of the x axes.

limY

Limits of the y axes.

dim_red

Dimensionality reduction to use for the visualization.

dot_size

Size of the dots.

apha

Alpha of the dots.

color_discrete

Colors for discrete parameters.

color_gradient

Colors for continous parameters.

remove_guide

Logical, if you want to remove the guide.

facet_by_variable

Logical if the plot should be splitted by the categorical variable used.

label_clusters

Logical: If clusters should be labeled with a text box.

label_size

Size of the labels.

label_color

Color of the labels.

raster

TRUE or FALSE, if plot should be returned as raster image

Value

plot marker or list of markers

Details

plot_marker