Heatmap of scaled expression by cell population

plot_marker_HM() generates a heatmap of scaled mean marker expression for each cell population.

Usage

plot_marker_HM(
  fcd,
  expr_slot = "orig",
  marker_to_exclude = NULL,
  cluster_slot,
  cluster_var,
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  maxvalue = NULL,
  size = 10,
  title = "Heatmap of scaled expression"
)

Arguments

fcd

flow cytometry data set, that has been subjected to clustering or cell type label prediction with cyCONDOR

expr_slot

expr_slot from which to take marker expression values, default is "orig". Corrected input data should be handled cautiously.

marker_to_exclude

(optional) vector of characters indicating which features in expression matrix should not be included in the heatmap.

cluster_slot

string specifying which clustering slot to use to find variable specified in cluster_var

cluster_var

string specifying variable name in cluster_slot that identifies cell population labels to be used (e.g. clusters, metaclusters or predicted labels).

cluster_rows

logical indicating if heatmap rows should be clustered (default: FALSE).

cluster_cols

logical indicating if heatmap columns should be clustered (default: FALSE).

maxvalue

max value for the coloring (default: NULL, automatically defined).

size

size of the individual squares and font.

title

character string, title of the plot

Value

A heatmap of scaled mean expression, depicting markers in rows and cell populations in columns.

Details

plot_marker_HM