violin plot of marker expression in cell populations

plot_marker_violinplot() plots the expression of selected markers as violin plots for cell populations.

Usage

plot_marker_violinplot(
  fcd,
  marker,
  expr_slot = "orig",
  cluster_slot,
  cluster_var,
  cluster_to_show = NULL,
  group_var = NULL,
  color_palette = cluster_palette
)

Arguments

fcd

flow cytometry data set, that has been subjected to clustering or cell type label prediction with cyCONDOR

marker

vector of characters indicating which features of the expression matrix should be plotted.

expr_slot

expr_slot from which to take marker expression values, default is "orig". Corrected input data should be handled cautiously.

cluster_slot

string specifying which clustering slot to use to find variable specified in cluster_var

cluster_var

string specifying variable name in cluster_slot that identifies cell population labels to be used (e.g. clusters, metaclusters or predicted labels).

cluster_to_show

vector of strings indicating levels in cluster_var that should be included for plotting.

group_var

(optional) string indicating variable name in cell_anno that should be used to split violin plots.

color_palette

vector of colors to be used to fill violin plots, when group_var is used

Value

plot_marker_violinplot() returns either one plot in case only one marker is provided via marker argument or a list of plots, if several markers are requested.

Details

plot_marker_violinplot

The violin plots are plotted with default parameters of ggplot2's geom_violin and horizontal lines indicate the median.