Read FlowJo Workspace

Loading and transforming the data to create a flow cytometry dataset from a Gate Set object for the analysis with the cyCONDOR workflow.

prep_fjw takes a gate set object as input and returns a fcd with the FlowJo gating information saved in fcd$anno$cell_anno.

Usage

prep_fjw(
  data_gs,
  inverse.transform = FALSE,
  transformation = NULL,
  remove_param = NULL,
  merge_anno = FALSE,
  anno_table = NULL,
  separator_anno = ",",
  filename_col = NULL
)

Arguments

data_gs

Gate Set object, e.g. created by using open_flowjo_xml and flowjo_to_gatingset from the CytoML package.

inverse.transform

Logical: if the data should be reverse transformed or kept with FlowJo transformation (default = FALSE).

transformation

If inverse.transform = TRUE, type of new transformation to perform. Select one of the following: "auto_log" (autologicle, recommended for flow cytometry data), "arcsinh" (arcsinh transformation), "clr" (centered-log-ratio) or "none" (no transformation).

remove_param

Parameters to be removed from the fcd.

merge_anno

Logical: If sample anno should be merged to the fcd.

anno_table

If merge_anno = TRUE, path to the annotation table text file. The annotation table should contain one column with the file names of all .fcs or .csv files to read in and optionally additional columns with further sample information (e.g. "sample_id", "condition").

separator_anno

Separator used in the annotation file, by default separator_anno = ",".

filename_col

Name of the column of the anno_table containing the file name matching with the .fcs files.

Value

An object of class 'flow cytometry dataframe' (fcd) is returned.

read_flowjo_workspace

Details

Read FlowJo workspace