runDM

Run Diffusion Map dimensionality reduction.

Usage

runDM(
  fcd,
  input_type,
  data_slot,
  k = 10,
  seed = 91,
  prefix = NULL,
  nPC = ncol(fcd[[input_type]][[data_slot]]),
  markers = colnames(fcd$expr[[data_slot]]),
  discard = FALSE
)

Arguments

fcd

flow cytometry dataset.

input_type

data to use for the calculation, e.g. "expr" or "pca".

data_slot

data slot to use for the calculation, e.g. "orig" or "norm".

k

K used for the analysis. Default = 10.

seed

A seed is set for reproducibility.

prefix

Optional prefix for the slot name of the output.

nPC

Number of principal components to use for the analysis. Default = All.

markers

Vector of marker names to include or exclude from DM calculation according to the discard parameter. See functions used_markers and measured_markers for the extraction of markers directly from the condor object.

discard

LOGICAL if the markers specified should be included, "False", or excluded, "True", from the DM calculation. Default = F.

Value

The function returns a fcd including a data frame with the DM coordinates saved in fcd$diffmap. The name of the output consists of the prefix (if given) and the data slot. If a nPC is given it will be added to the output name.

Details

runDM

See [Philipp Angerer et al. (2015): destiny: diffusion maps for large-scale single-cell data in R. Helmholtz-Zentrum München.] (http://bioinformatics.oxfordjournals.org/content/32/8/1241)