train_cytonorm

Wrapper function around CytoNorm.train from the CytoNorm package.

Usage

train_cytonorm(
  fcd,
  batch_var,
  remove_param = NULL,
  seed = 91,
  files = NULL,
  data_path = NULL,
  FlowSOM_param = list(nCells = 5000, xdim = 5, ydim = 5, nClus = 10, scale = FALSE)
)

Arguments

fcd

flow cytometry dataset

batch_var

Column name of batch variable from fcd$anno$cell_anno.

remove_param

Parameters/markers which should be excluded for learning the batch effect and training the model.

seed

A seed is set for reproducibility.

files

Vector of FCS file names of reference samples which are used for training the model. If files == NULL, all files contained in the fcd are used.

data_path

File path to folder where .fcs files contained in the fcd are stored. This parameter does not need to be provided, unless the folder where the .fcs files are stored has changed.

FlowSOM_param

A list of parameters to pass to the FlowSOM algorithm. Default= list(nCells = 5000, xdim = 5, ydim = 5, nClus = 10, scale= FALSE)

Value

The function returns a fcd with the trained model saved in fcd$extras$cytonorm_model.

Details

train_cytonorm

train_cytonorm' takes a fcd as an input and learns the batch effect of a given batch variable across reference samples provided by the user using the CytoNorm algorithm. This function returns a fcd with the trained model which can be used as input for the run_cytonorm function to normalize samples with the trained model. See [Van Gassen et al., 2019](https://doi.org/10.1002/cyto.a.23904) for more details on CytoNorm.